+++ clang/include/clang/Serialization/ASTBitCodes.h
@@ -1759,6 +1759,9 @@ enum StmtCode {
   /// A ConvertVectorExpr record.
   EXPR_CONVERT_VECTOR,
 
+  /// A ConvertFromArbitraryFPExpr record.
+  EXPR_CONVERT_FROM_ARBITRARY_FP,
+
   /// BlockExpr
   EXPR_BLOCK,