+++ clang/lib/AST/ItaniumMangle.cpp
@@ -5968,6 +5968,17 @@ recurse:
     Out << "v18co_yield";
     mangleExpression(cast<CoawaitExpr>(E)->getOperand());
     break;
+
+  case Expr::ConvertFromArbitraryFPExprClass: {
+    const auto *CE = cast<ConvertFromArbitraryFPExpr>(E);
+    NotPrimaryExpr();
+    mangleVendorType("__builtin_convert_from_arbitrary_fp");
+    mangleTemplateArgExpr(CE->getSrcExpr());
+    mangleVendorType(CE->getFormat());
+    mangleType(CE->getTypeSourceInfo()->getType());
+    Out << 'E';
+    break;
+  }
   case Expr::SYCLUniqueStableNameExprClass: {
     const auto *USN = cast<SYCLUniqueStableNameExpr>(E);
     NotPrimaryExpr();