+++ clang/lib/AST/StmtProfile.cpp
@@ -1656,6 +1656,12 @@ void StmtProfiler::VisitConvertVectorExpr(const ConvertVectorExpr *S) {
   VisitExpr(S);
 }
 
+void StmtProfiler::VisitConvertFromArbitraryFPExpr(
+    const ConvertFromArbitraryFPExpr *S) {
+  VisitExpr(S);
+  VisitType(S->getTypeSourceInfo()->getType());
+}
+
 void StmtProfiler::VisitChooseExpr(const ChooseExpr *S) {
   VisitExpr(S);
 }