+++ clang/lib/Sema/SemaExceptionSpec.cpp
@@ -1300,6 +1300,7 @@ CanThrowResult Sema::canThrow(const Stmt *S) {
   case Expr::ShuffleVectorExprClass:
   case Expr::StmtExprClass:
   case Expr::ConvertVectorExprClass:
+  case Expr::ConvertFromArbitraryFPExprClass:
   case Expr::VAArgExprClass:
   case Expr::CXXParenListInitExprClass:
   case Expr::CXXExpansionSelectExprClass: